Factor out tmdb store
