+++ web/src/api/v1/tmdb/types.ts
@@ -1,3 +1,20 @@
+export type TmdbConfiguration = {
+  change_keys?: string[];
+  images?: {
+    base_url: string;
+    secure_base_url: string;
+    backdrop_sizes: string[];
+    logo_sizes: string[];
+    poster_sizes: string[];
+    profile_sizes: string[];
+    still_sizes: string[];
+  }
+}
+
+export type TmdbState = {
+  configuration?: TmdbConfiguration
+}
+
 export type TmdbSearchResultCommon = {
   adult: boolean;
   backdrop_path: string;
