+++ web/src/pages/App.tsx
@@ -1,8 +1,6 @@
 import { AppContainer } from './App.styles'
 import { ButtonBar } from '@/components/ButtonBar'
-import { endpoints, type ApiModel } from '@/api/endpoints'
-import { useAppDispatch, useAppSelector } from '@/api'
-import { tmdbActions } from '@/api/v1/tmdb/store'
+import { useAppDispatch } from '@/api'
 import { StatusScroller } from '@/components/status/StatusScroller'
 import { TocGrid } from '@/components/toc/TocGrid'
 import { useGetErrorQuery } from '@/api/v1/error/api'
@@ -13,18 +11,9 @@ import { CombinedShowMovieForm } from '@/components/forms/ContentForm'
 function App() {
   const dispatch = useAppDispatch()
 
-  const tmdbConfiguration = useAppSelector((state) => state.tmdb.configuration)
   const errorData = useGetErrorQuery()
 
   dispatch(socketConnect())
-
-  if (!tmdbConfiguration) {
-    fetch(endpoints.tmdb.configuration())
-      .then(response => response.json() as Promise<ApiModel['v1']['tmdb/configuration']>)
-      .then(({ data }) => {
-        dispatch(tmdbActions.setTmdbConfiguration(data))
-      })
-  }
 
   return (
     <AppContainer>
